Migration log formatting to config; remove colorization from file log
